Moved exceptions to their own file, fixed schema tests.
