+++ db/records/exceptions.py
@@ -1,0 +1,10 @@
+from sqlalchemy_filters.exceptions import FieldNotFound
+
+
+# Grouping exceptions follow the sqlalchemy_filters exceptions patterns
+class BadGroupFormat(Exception):
+    pass
+
+
+class GroupFieldNotFound(FieldNotFound):
+    pass
