+++ mathesar/tests/test_models.py
@@ -11,7 +11,7 @@ def test_schema_name_sets_cache(monkeypatch, test_db_model):
         models.Schema, '_sa_engine', lambda x: None
     )
     monkeypatch.setattr(
-        models.schemas, 'get_schema_name_from_oid', lambda x, y: 'myname'
+        models.schema_utils, 'get_schema_name_from_oid', lambda x, y: 'myname'
     )
     cache.clear()
     schema = models.Schema(oid=123, database=test_db_model)
@@ -25,7 +25,7 @@ def test_schema_name_uses_cache(monkeypatch, test_db_model):
     )
     cache.clear()
     with patch.object(
-            models.schemas, 'get_schema_name_from_oid', return_value='myname'
+            models.schema_utils, 'get_schema_name_from_oid', return_value='myname'
     ) as mock_get_name:
         schema = models.Schema(oid=123, database=test_db_model)
         name_one = schema.name
@@ -43,7 +43,7 @@ def test_schema_name_handles_missing(monkeypatch, test_db_model):
     def mock_name_getter(oid, engine):
         raise TypeError
     monkeypatch.setattr(
-        models.schemas, 'get_schema_name_from_oid', mock_name_getter
+        models.schema_utils, 'get_schema_name_from_oid', mock_name_getter
     )
     schema = models.Schema(oid=123, database=test_db_model)
     name_ = schema.name
